Unseq AA impl
In this revision, we introduce a new alias-analysis algorithm, which we call `unseq-aa`, based on the must-not-alias relationships identified by our algorithm. At each initialization, `unseq-aa` makes a pass over the program and collects all must-not-alias relationships through the custom intrinsic calls present in the code.
It is at this point that LLVM's function summary database is queried and only those predicates are added whose associated functions are known to be `readnone`. `unseq-aa` then answers queries using these collected relationships.