AA Eval
In this revision, we add a command line option to enable or disable running the AliasAnalysisEvaluator pass. We also modify the `aa-eval` pass to collect statistics related to the must-not-alias predicates generated and the UnsequencedAliasAnalysis pass. For our evaluation, we count the number of times `unseq-aa` returns a must-not-alias result when all
other alias analysis algorithms together return may-alias. We also collect other statistics related to the percentage of the must-not-alias predicates generated which turn out to be used to make aliasing related decisions.