Intrinsic addition
In this revision, we add a custom variadic LLVM intrinsic called `llvm.unseq.noalias` whose first argument represents the serial number, the next two arguments represent the LValues corresponding to the subexpressions that must not alias, and the remaining arguments are the corresponding function calls that tag this predicate.

The LValue arguments of our custom intrinsic are wrapped inside `metadata nodes` that ensures that these references to LValues are only external and are not part of the actual program computation. The custom LLVM intrinsic returns `void` and is marked `readnone` (indicating that it is a pure function), and thus provides a convenient method to maintain extra information in the program without hampering any optimization opportunity. Further, because the arguments to the intrinsic calls are LLVM Values wrapped in metadata nodes, the arguments get transformed along with the transformations to the rest of the LValues, as optimizations proceed.

We modify LLVM's dead-code elimination pass to not eliminate calls to our custom intrinsic until the very last step of code generation. We also ensure that the extra instructions due to our intrinsic calls are not counted while assessing the size of basic blocks, etc. (e.g., for inlining or loop unrolling decisions).