CodeGenImpl
In this revision, we make additions to LLVM's CodeGeneratorAction to lower the predicates inferred in the earlier steps to LLVM IR. To lower a predicate, it will require a map from every (sub)expression to its corresponding LLVM (temporary) register name, that represents its "L Value". Since code generation happens in a bottom up manner for expressions, following a deterministic evaluation order (left-to-right), the expression to LValue map is filled in the same order.
At the end of the code generation for an expression, all its sub-expressions have been assigned LValues, which means all the members of a must-not-alias predicate tuple associated with it have mappings. The predicates are "lowered" and  emitted as intrinsics, as described earlier, with arguments being the LValues, right after the code for the expression is emitted.

Because LLVM uses SSA form for all scalar register values (and an expression LValue is always scalar), we can associate the must-not-alias property with the expression LValues for the entire program.